+++ week2/Assignment1/src/components/mainStatic.js
