+++ week2/Assignment1/src/components/renderFunction.js
